ECS-368: Validate component set package size
Restrictions are added on the number of allowed MB's and the number files in the component set and specifcally the custom data folder. We don't want a massive component set to be uploaded to the ECS backend as it takes up unnecessary storage and it takes too long to process (especially when there are a lot of files to unpack).