CSH-8382: roll back 523b8ba and solve in a different way
Fixes the problem in ecs backend where the filePaths and getFileContent paths no longer match because one is normalized and the other isn't. Also keeps the getFileContent, getFileSize and filePaths arguments of the validate function consistent: all take normalized paths. The problem of having a path.normalize in the browser is solved by separating the loading of the renditions is a separate function that is not imported in the browser when using parseDefinition. Additionally adds deepFreeze to ensure the components definition is not altered while parsing it into a component set.