CSH-8707: introduce object data type.
This is used to indicate the data refers to an object in Studio.
When a property is created with this data type, it ensures placed relations are created with that object.
For now it can only be used with the experimental "object-select" UI control type.
Archiving to assets is not supported (the image will not be picked up).